update rad test 46
